requirements.txt
streamlit
azure-ai-inference
fpdf
PyPDF2
